config.exs: Add configure options from environment variables
Add the possibility to configure URL and network_id to an aeternity node
from environment variables. This makes ae-channel-service easier to
configure when deployed, for example, to a VM or docker.

The config uses System.get_env/2 to provide default values when these
values are not configured through environment variables or explicitly
set in config.exs. System.get_env/2 requires Elixir version >= 1.9.0.